Include help for Add Game to show how 4 and 5 player games are entered
This commit adds some help text at the bottom of the Add Game screen to
explain how to change to 5-player mode and back to 4-player mode. It
also fixes a bug where the total score was not updated immediately after
deleting the 5th player.  The help text is updated dynamically depending
on the current mode.

This change also adds a thousands separator comma to the point total.