Fix use of map in seating.py to avoid problem in Python 3
The map function behaves slightly differently in Python 2 vs 3.  In
Python 2, it returns a sequence, but in Python 3 it can return an map
object/iterator whose length can't be measured and whose elements
can't be sliced.  This change wraps the output of map with a list
function so the rows fetched from the db are processed like a list.