Fix cookie handling for Python3 so admin and stylesheet cookies work
The tornado module returns cookie values as byte arrays.  In Python3
byte arrays do not match equivalent strings using the == operator.
This change wraps the calls to get_secure_cookie in a stringify
utility that forces the cookie value to a unicode string if it is not
None or already in string form.  It also also logger for "WebServer"
which logs user log in and log out actions.