Make Add Players from Meetup button contingent on meetup configuration
For clubs that don't use meetup.com, there's no need for the button
that adds the RSVPs.  This change checks whether the API parameters are
configured or not and only displays the button if they are.  It also
adds today's date to the button label to make it clearer what RSVPs
will be added.