Change leaderboard to shift filtering and numbering into browser javascript
This changes the Ajax callback for the leaderboard data to fetch the
data for a particular kind of period and puts the filtering by minimum
number of games in the leaderboard.js.  The numbering of the table
rows can either be "PLACE" for the overall ordering of the score in
that period or "ORDER" for the ordering among just the visible rows.
Gaps in the overall ordering due to filtering are indicated with
ellipses.  Clicking on any ordering column header toggles between the
2 modes.