Templated playerstats and added stats for different periods
Changed playerstats to have data handler that returns the stats data in
JSON to use the Mustache templater to fill in each table of stats.

Changed behavior of POST to playerstats to not do a redirect when values
haven't changed.

Added some utility functions for managing quarter names and computing
previous and next quarters.