Use CLUBNAME from settings.py in all messages
This change updates the templates and email content to use the
CLUBNAME parameter in settings.py for the club name.  The
expiration of invites and password reset links is now a parameter
in settings.py.

This commit changes the invite and verify mechanisms to avoid
extending invitations for or creating duplicate accounts.  When
verifying invites, expired links get a different message from
duplicates, and both conditions delete the link ID.  There were
also a few corrections to the wording of the email messages.

Changed use of rowcount in sqlite cursor object since it doesn't really
reflect the count of the rows of the query that was executed.

Commented out references to the non-existant admin.css and login.css files
(but still get file not found errors in the Tornado logs).