Added pie chart & legend to show rank distribution in player stats
Moved Date range and Games played to top of stats table so all rank
related stats are together in bottom rows.

Changed structure of rank histogram in JSON to list of dicts with
'rank' and 'count' to be easier for processing in D3.

Added color definitions for all ranks to the 3 color preference files.