Enable some forms of auto-database updates for schema changes
This commit abstracts the database schema into a dictionary of tables with
lists of field specifications.  When the db is init'd it checks if the
actual database has all the fields in the spec and alters the database tables
if it doesn't.  It has an option to force the changes or to prompt the user
for individual changes.