Change Add Quarter to Manage Quarter Settings and improve db schema update
This commit changes how quarters are managed.  Instead of adding quarters
to set the gamecount, it presents all possible quarters based on values
in the Scores table and allows admins to edit or delete the corresponding
records in the Quarters table.

The message.html template is expanded to include a "next" link for the
expected next URL to visit after reading the message.

The main.py program now accepts a command line argument, -f or --force, to
force schema changes without prompting the user.  The mix of locally defined
command line arguments and those inherited from tornado are displayed
when users specify -h or --help on the command line.

It also introduces schema changes to the Quarters table to avoid duplicate
entries and to add a setting for an UnusedPointsIncrement value.