Allow editing of UnusedPointsIncrement
This setting can be set per quarter and will control whether or not
scores can be entered with less than the number of players times 25,000.
If it is non-zero, unused points can be entered as long as they are
multiples of this increment.  For example, this can be used to handle
when unclaimed riichi sticks remain after game play ends.