Added ability to store and display unused points with every game
This commit changes the user inteface to enable the entry of unused
points when recording scores and the display of those unused points
in game history and user history.

The addgame handler is moved from main.py to its own file, addgame.py,
and the addgame.js file is removed by placing its work in gameeditor.js.
It fixes a bug with the updating of the help information when adding
and removing a 5th player position.

It adds a sanity check that all the players names/ids are distinct when
entering scores.