Added color highlights to unused points entry to indicate good and bad values
In addition to GUI highlights, gameeditor.gameComplete and db.addGame
verify that the unused points entry is a multiple of the unused points
increment setting.

Simplified color CSS specifications for pie chart and its legend by
adding common class names in addition to the rank-specific ones.

Fixed a typo in the playerstats.html template.