Update beautify.sh to accept specific files on command line
Allow beautify.sh to operate on specific files or every javascript
file in the static/js directory.  Also change the command line
syntax so it works with both the python and node.js versions of
js-beautify.  NOTE: this requires dropping the -p argument to
js-beautify which is not documented in the command line help and
is only accepted by the node.js version of js-beautify.  I'm not
sure what it is supposed to do.