Implement proper handling of tie scores and apply chombos before ranking
The way rankings for each game are computed is changed to average
the uma bonuses when tie scores occur and to include chombo
penalties in the score before ranking.  The initial score per
player and the chombo penalty amount are now configurable
parameters.  The parameters in defaults.py are reorganized into
more functional categories.  References to the initial score
per player in various parts of the code use the parameter value.

This also fixes a bug where unused points were not being shown in
game history displays and the games are sorted by date and game
ID in decreasing order.