Show user name when logged in
This commit changes the NavBar to show the email of the logged in user
in place of the Settings button.  The user name is shortened to be
just the email account name without the domain, or the shortest
unambiguous email substring among all users.  The user name is
hyperlinked to the Settings.

I adjusted the CSS to try to place the user name right justified in
the last row of the navbar, but couldn't figure out how to place it
there and keep the other buttons of the navbar centered.  I avoided
using absolute positioning since I want the navbar to wrap into
multiple rows on small displays.

Since handler.py now imports db.py, I had to remove the import
leaderboard from db.py since that caused a import loop where
leaderboard expected handler to already be loaded.

I fixed some missing uri definitions in the calls to render the
login.html page.