Enable setting and editing end-of-quarter tournament qualifying criteria
This change adds fields to the Quarters records to track not only the
number of games needed to drop a score from the average, but also the
number of games or distinct days played in order to qualify for the
end of quarter tournament.  Default values for these fields are taken
from the defaults.py/mysettings.py file, but are editable for each
quarter.  The quarter editor fills in values from the most recent
quarter record instead of the default settings when editing a missing
record.

The deletequarter.html template is not being used so I deleted it.