Added Players display/management screen and way to create future quarters
Added a new link from the home page to a Players display.  The Players
page lists all players with recorded scores, their meetup name, and
the quarters for which they are members.  For unprivileged users, they
can only see the information and follow links to a player's stats
page.  For admin users, they are presented inputs to edit the
player attributes and checkboxes to control their membership status
per quarter. The page has an empty javascript file TBD, to make it
correctly show/hide a menu controlling the subset of available quarters
that are visible and make updates as the information is changed by admins.

Since memberships reference quarters in the future, the Manage Quarters page
now will list a few more quarter records that can be edited to fill in
settings.  This is a pre-req for being able to set player membership status
for those quarters.  The number of future quarters is controled by a new
setting in defaults.py.