Manage Quarters now shows all Quarters. Enabled toggle of Qtrs in Players
This partially enables interactivity in the Players dialog to show and hide
the menu that controls which quarters are visible, along with correctly
toggling the visibility when controls are clicked.  It introduces a new
setting for the number of Membership quarters to show initially in the Players
dialog.

Since Players needs the creation of future Quarters records, the Manage
Quarters dialog now finds the union of all quarters mentioned in a Scores
record with all the existing Quarters records.

Updates to Player attributes and membership are not yet making changes to
the database.