Moved location and name for Membership Quarters visibility menu
The control for popping up the Membership Quarters visibility menu is
moved to the column label over the list of quarters.  The popup menu
now is dismissed if clicks are made on any other input outside of the
menu.

Also ran beautify.sh with the node.js version of js-beautify to standardize
the javascript code.