Fix Membership Quarters Visualization control to not auto-dismiss
Moved the table that implements the visualization control to the preceding
cell in the table so that it's not under the DOM for the label that pops it
up.  That prevents it from be toggled off on the first click to hide/expose
a quarter column.  It also moves the popup menu to cover the Meetup Name
column which is better.