Add DateCount to Leaderboards
This commit adds a field to each leaderboard row showing the number of
distinct dates of games played in that period.  When upgrading an
existing database, the new field is only populated when the
leaderboard for that period is recalculated (i.e. when a game is added
or edited in the period).  The field is displayed in a new column..