Display membership and end-of-quarter tournament eligibility in Leaderboards
This change adds information to the Quarterly Leaderboards to show
whether the player was a member that quarter, and if they played enough
to be eligible for the end-of-quarter tournament.  Membership is indicated
with an asterisk next to the player's name.  Eligibility is indicated with
italic font for the player's row and a change in background color.  A legend
for these indicators is added to the bottom of the page.  The legend and
indicators are hidden for bi-annual and annual leaderboards.

The way that leaderboards are generated was changed, too.  Instead of
looping over several queries with different numbers of dropped games,
a single query to get the basically quarterly information is run and then
separate queries for each player that qualified for dropping games is
run to get their dropped game average.