Added Quarterly Pariticipation History to Player Statistics page
At the bottom of the Playerstats page, a quarterly participation history
lists all of the quarters in the system and whether the individual played
games that quarter, was a member, and was eligible for the end of quarter
tournament.  The style is the same as for the Leaderboard display.

Fixed the get_eligible function to correctly find all quarters in the
system.