Expanded Quarterly Participation History into Timeline showing future quarters
The Players Statistics page now shows a Quarterly Participation Timeline
that includes any quarters that have been registered but do not yet have
scores recorded.  This allows players to see that they are marked as members
for future quarters (on the stats page rather than going to the Players list).

Added a setting for the maximum length of the timeline.