Added command line tool to measure time while updating leaderboards
This commit makes the leaderboard.py module a command line tool as well
that can be used to recalculate leaderboards.  It uses the the Python
timeit package to run the calculation multiple times to determine how
long it takes.

I also separated out the formatting of the SQL statement from its
execution to help in debugging.