Allow editing of Quarters without dropping quarter record
In the add-membership-status branch, the Memberships table has a
foreign key constraint on the Quarters table.  Once you've made
a scores database with this constraint, attempts to edit a quarter
record fail when the Admin page tried to delete an insert a new
record.  This change makes it an update/replace operation.  The
update/replace works for both the master and add-membership-status
branches.