Changed seater to include non-seated players
The Seater now pulls in all players from Meetup, regardless of
whether they are known to the database, creating Player records
as needed.  It puts these players in a new status, which has them
showing up in the seater list of players, but doesn't seat them
at a table.  Clicking the status icon now cycles through the
states of Unseated -> Prioritized for 4 player table ->
Unprioritized -> Unseated.  The list of players is displayed
in order with all the unseated players at the bottom and
ordered by name otherwise.  This has the side-effect of moving
players within the list when they transition from seated to
unseated or vice-versa.  That may not be desirable to all
users.  Unseated players can remain in the database without
having any scores, so I disabled the periodic cleanup of
those player records.  If we do need to clean them up, it
should be done much less frequently than once an hour.

Added hyperlinks to playerstats from seater page.  This is useful
for admins to edit the name or meetup name of a new player and
for end users who want to see the stats of their opponents.
Added an auto-refresh checkbox to seater page, checked by
default.  It's probably more useful to developers than to end
users so it's at the bottom.  Rearranged some of the controls on
the seater and playerstats pages to make them more compact and
readable/reachable-by-scrolling on small screens.

Made PlayerStats work when a player has no past scores recorded, so that
the player's account info could still be edited.  Fixed bug in post
method if quarter was included in URI.

Added feature to playerstatsdata to interpret 'latest' to mean
the latest quarter that the player has scores. Changed names of
quarterly stats to be of the form "2018 2nd Quarter". Fixed bug
in playerstatsdata if player had no scores; it now returns a reasonable
json response.

Fixed colors for seating display in diffent pallettes.