Load external files from static in developer mode
This change stores copies of the external Javascript and CSS files
used by the codebase in the appropriate static/js or /css directory.
When settings.DEVELOPERMODE is True, the HTML files load the JS and
CSS files from the static location.  This makes developing faster
when you have a slow connection to the Internet, and enables self-
contained operation when there is no Internet connection.