Fix Add Game and Game Editor problems
This change ensures that the number of players and any unused points
are properly populated into the Edit Game dialog. The Add Game dialog
properly adds and removes a 5th player along with a new constraint
that the fifth player name be empty for removal. This is to enable the
(rare) condition that a fifth player with an exact score of 0 occurs.
It fixes an issue where the player name clear button for player
completion was not being created for the 5th player in Add Game.
When player names are altered, the check to remove the 5th player
and the submit button enable/disable status is updated.